// src/common/enums/order.enums.ts
export enum OrderStatus {
    NEW = 'NEW',
    FILLED = 'FILLED',
    REJECTED = 'REJECTED',
    CANCELLED = 'CANCELLED',
  }
  
  export enum OrderType {
    MARKET = 'MARKET',
    LIMIT = 'LIMIT',
  }
  
  export enum OrderSide {
    BUY = 'BUY',
    SELL = 'SELL',
    CASH_IN = "CASH_IN",
    CASH_OUT = "CASH_OUT"
  }